Bump lock tokens gas due to Boron chain upgrade (#140)
The upgrade did a patch to create fewer locks. However there was a regression in gas amounts for folks who already have many locks pre-upgrade. We didn't catch this, really sorry :/

(We wrote a test for testing this ... just we tested the Keeper LockTokens function, not the `LockTokens` message. Hence we missed this)

Testing on mainnet, I get ~130k gas needed, so 400k should be more than enough for people with way too high gas params