Use timeout_height from signed amino doc
In institutional cases, the signature process may take a while.

This is why we override:

```
  sequence: signedSequence
```

Since maybe the sequence number has changes since the dApp sent its
requested sign doc to the wallet - the wallet may update the
sequence number to a new value, and the dApp should use the returned
value - which was signed on.

The same reasoning should follow for `timeout_height` which is a newer
addition to the protocol than sequence number.