feat: dust to osmo swap flow implementation
Reusing the existing infrastructure for swaps encoded in the useSwap
seemed to be the favorable direction.

The main trick to it is to wrap the hooks into lightweight components
and handle the batching through the component interactions.

(implmeneting `useDustToOsmo` from scratch would have been possible but
it would have needed more testing and it would have duplicated at least
some of the existing functionality — eg handling 1 click trading)