refactor: rework mobx <> 1CT session reactivity to be push based
The useSwap hook was not updating instantly after the create1CTSession
mutation, but before the onSuccess callback. This caused the next swap
transaction to either fail or not get signed properly through 1CT.

This could be worked around with waiting for the changes to propagate,
but that solution can be error prone.

To implement a proper fix, the logic for handling updates has been moved
into the mobx store. The `useOneClickTradingSession` now uses mobx’s
reaction() to react immediately to changes and update the state as soon
as it changes. This ensures that the session info is always available
and up to date after change — eg after create1CT mutation but before the
`onSuccess` callback.

The session data is stored as regular state, so any components using
this hook can rely on react’s built-in reactivity without needing to
work with mobx observables directly.

Finally, a useEffect is used to clean up resources when the component
unmounts.